renamed app to config
