+++ pkg/config/config_load.go
@@ -1,4 +1,4 @@
-package app
+package config
 
 import (
 	"io"
