+++ pkg/config/config_load_test.go
@@ -1,4 +1,4 @@
-package app
+package config
 
 import (
 	"os"
