+++ pkg/config/paths.go
@@ -1,4 +1,4 @@
-package app
+package config
 
 // Paths contains the paths to directories used by the application
 type Paths struct {
