+++ pkg/config/validator.go
@@ -1,4 +1,4 @@
-package app
+package config
 
 import (
 	"errors"
