+++ pkg/config/validator_test.go
@@ -1,4 +1,4 @@
-package app
+package config
 
 import "testing"
