+++ pkg/templating/pattern.go
@@ -5,11 +5,11 @@ import (
 	"html/template"
 	"strings"
 
-	"github.com/matthewchivers/journal/pkg/config/app"
+	"github.com/matthewchivers/journal/pkg/config"
 )
 
 // ParsePattern creates a new path for a journal entry based on a path template
-func ParsePattern(rawTemplate string, entry app.Entry) (string, error) {
+func ParsePattern(rawTemplate string, entry config.Entry) (string, error) {
 	data, err := PrepareTemplateData(entry, strings.Contains(rawTemplate, "{{.WeekCommencing}}"))
 	if err != nil {
 		return "", err
