+++ pkg/templating/pattern_test.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/matthewchivers/journal/pkg/caltools"
-	"github.com/matthewchivers/journal/pkg/config/app"
+	"github.com/matthewchivers/journal/pkg/config"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -25,56 +25,56 @@ func TestParsePattern(t *testing.T) {
 	tests := []struct {
 		name         string
 		pathTemplate string
-		entry        app.Entry
+		entry        config.Entry
 		want         string
 		wantErr      bool
 	}{
 		{
 			name:         "Valid template: year/month/day/",
 			pathTemplate: "{{.Year}}/{{.Month}}/{{.Day}}/",
-			entry:        app.Entry{},
+			entry:        config.Entry{},
 			want:         currentYear + "/" + currentMonth + "/" + currentDay + "/",
 			wantErr:      false,
 		},
 		{
 			name:         "Valid template: year/month/week/entryID/",
 			pathTemplate: "{{.Year}}/{{.Month}}/{{.WeekNumber}}/{{.EntryID}}/",
-			entry:        app.Entry{ID: "foo-note"},
+			entry:        config.Entry{ID: "foo-note"},
 			want:         currentYear + "/" + currentMonth + "/" + string(rune(currentWeek)) + "/foo-note/",
 			wantErr:      false,
 		},
 		{
 			name:         "Valid template: year/month/wc weekCommencing/weekday/entryName/",
 			pathTemplate: "{{.Year}}/{{.Month}}/wc {{.WeekCommencing}}/{{.WeekdayName}}/{{.EntryID}}/",
-			entry:        app.Entry{ID: "foo-note"},
+			entry:        config.Entry{ID: "foo-note"},
 			want:         wcCurrentYear + "/" + wcCurrentMonth + "/wc " + weekCommencing.Format("2006-01-02") + "/" + currentWeekdayName + "/foo-note/",
 			wantErr:      false,
 		},
 		{
 			name:         "Valid template: year/month/week-commencing/template/",
 			pathTemplate: "{{.Year}}/{{.Month}}/{{.WeekCommencing}}/{{.EntryID}}/",
-			entry:        app.Entry{ID: "foo-note"},
+			entry:        config.Entry{ID: "foo-note"},
 			want:         currentYear + "/" + currentMonth + "/" + weekCommencing.Format("2006-01-02") + "/foo-note/",
 			wantErr:      false,
 		},
 		{
 			name:         "Path template without placeholders",
 			pathTemplate: "static-path/foo-note.md",
-			entry:        app.Entry{ID: "foo-note"},
+			entry:        config.Entry{ID: "foo-note"},
 			want:         "static-path/foo-note.md",
 			wantErr:      false,
 		},
 		{
 			name:         "Invalid path template",
 			pathTemplate: "{{.Year}/{{.Month}}/{{.Day}}/{{.EntryID}}/", // Missing a closing brace
-			entry:        app.Entry{ID: "foo-note"},
+			entry:        config.Entry{ID: "foo-note"},
 			want:         "",
 			wantErr:      true,
 		},
 		{
 			name:         "Empty path template",
 			pathTemplate: "",
-			entry:        app.Entry{},
+			entry:        config.Entry{},
 			want:         "",
 			wantErr:      false,
 		},
